+++ CDTDatastore/query/CDTQIndexManager.h
@@ -3,6 +3,7 @@
 //
 //  Created by Mike Rhodes on 2014-09-27
 //  Copyright (c) 2014 Cloudant. All rights reserved.
+//  Copyright © 2018 IBM Corporation. All rights reserved.
 //
 //  Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
 //  except in compliance with the License. You may obtain a copy of the License at
@@ -85,7 +86,7 @@ typedef NS_ENUM(NSInteger, CDTQQueryError) {
  */
 @interface CDTQIndexManager : NSObject
 
-@property (nonatomic, strong) CDTDatastore *datastore;
+@property (nonatomic, weak) CDTDatastore *datastore;
 @property (nonatomic, strong) FMDatabaseQueue *database;
 @property (nonatomic, readonly, getter = isTextSearchEnabled) BOOL textSearchEnabled;