+++ CDTDatastore/touchdb/TD_Database.m
@@ -8,7 +8,7 @@
 // Modified by Michael Rhodes, 2013
 // Copyright (c) 2013 Cloudant, Inc. All rights reserved.
 //
-// Copyright © 2017 IBM Corporation. All rights reserved.
+// Copyright © 2017, 2018 IBM Corporation. All rights reserved.
 //
 //  Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
 //  except in compliance with the License. You may obtain a copy of the License at
@@ -642,6 +642,7 @@ static BOOL removeItemIfExists(NSString* path, NSError** outError)
 
 - (void)dealloc
 {
+    NSLog(@"dealloc td_database");
     if (self.isOpen) {
         // Warn(@"%@ dealloced without being closed first!", self);
         [self close];